+++ app/src/main/java/com/example/bookofrecipes/fragments/HomeAdvancedFragment.kt
@@ -14,7 +14,7 @@ import com.example.bookofrecipes.R
 import com.example.bookofrecipes.databinding.FragmentHomeAdvancedBinding
 import com.example.bookofrecipes.recyclers.recipes.RecipeRepository
 import com.example.bookofrecipes.recyclers.recipes.recyclers.ingredients.IngredientAdapter
-import com.example.bookofrecipes.recyclers.recipes.recyclers.steps.StepAdapter
+import com.example.bookofrecipes.recyclers.steps.StepAdapter
 
 class HomeAdvancedFragment : Fragment(R.layout.fragment_home_advanced) {
     private var binding: FragmentHomeAdvancedBinding? = null
